% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{abind_list}
\alias{abind_list}
\title{Combine multi-dimensional arrays with lists}
\usage{
abind_list(..., rev.along = NULL)
}
\arguments{
\item{...}{any number of lists with dimensions or a single list of list arrrays}

\item{rev.along}{dimension to reverse before binding (passed to \code{\link[abind]{abind}})}
}
\description{
Combine multi-dimensional arrays with lists
}
