% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{collapse_na_array}
\alias{collapse_na_array}
\title{Collapse blocks of NAs in an array with numeric dimension names}
\usage{
collapse_na_array(x)
}
\arguments{
\item{x}{array with names having numeric positional values}
}
\description{
Collapse blocks of NAs in an array with numeric dimension names
}
