% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{comb_vec_to_param_array}
\alias{comb_vec_to_param_array}
\title{Determine array of destination parameters from a source vector}
\usage{
comb_vec_to_param_array(comb_vec, comb_array, param_array = NULL,
  na_only = FALSE)
}
\arguments{
\item{comb_array}{array of 2xN data frames with mapping between vector and array}

\item{param_array}{starting array of destination parameters}

\item{na_only}{only overwrite values in param_array if they are NA}

\item{vector}{of source parameters}
}
\description{
Determine array of destination parameters from a source vector
}
