% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{coupling_param_idx}
\alias{coupling_param_idx}
\title{Get a list of logical arrays indicating which parameters correspond to scalar couplings}
\usage{
coupling_param_idx(param_list, comb_idx_offset = 0)
}
\arguments{
\item{param_list}{parameter list}

\item{comb_idx_offset}{offset for combination index mapping}
}
\value{
A named `list` of logical index arrays/vectors (matching
  `param_list$group_list`) that identify scalar-coupling parameters in
  `omega0_comb`.
}
\description{
Get a list of logical arrays indicating which parameters correspond to scalar couplings
}
