% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{extract_params}
\alias{extract_params}
\title{Extract parameters from fit object for use with make_fit_input}
\usage{
extract_params(fit, expand = 0)
}
\description{
Extract parameters from fit object for use with make_fit_input
}
