% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{fit_footprint}
\alias{fit_footprint}
\title{Determine the region of a spectrum containing the majority of the fit peaks}
\usage{
fit_footprint(fit, frac = 0.12)
}
\arguments{
\item{fit}{fit_output structure}

\item{frac}{fraction of total intensity used to define the footprint}
}
\description{
Determine the region of a spectrum containing the majority of the fit peaks
}
