% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{fit_peak_cluster}
\alias{fit_peak_cluster}
\title{Fit a cluster nearby peaks starting from a seed table of chemical shifts}
\usage{
fit_peak_cluster(
  spec_list,
  cs_start,
  spec_ord,
  f_alpha_thresh = 0.001,
  omega0_plus = c(0.075, 0.75),
  r2_start = 5,
  r2_bounds = c(0.5, 20),
  sc_start = NULL,
  sc_bounds = c(0, Inf),
  plot_main_prefix = NULL,
  peak_num_offset = 0,
  plot_fit_stages = FALSE
)
}
\arguments{
\item{spec_list}{list of spectra}

\item{cs_start}{matrix of starting chemical shifts}

\item{spec_ord}{order of spectral dimensions}

\item{f_alpha_thresh}{alpha threshold for F-test}

\item{omega0_plus}{omega0 bounds for each dimension}

\item{r2_start}{starting r2 values}

\item{r2_bounds}{numeric vector of length 2 with r2 bounds}

\item{sc_start}{starting scalar coupling values}

\item{sc_bounds}{numeric vector of length 2 with scalar coupling bounds}

\item{plot_main_prefix}{optional prefix for plot titles}

\item{peak_num_offset}{offset added to peak numbers in plots}

\item{plot_fit_stages}{logical indicating whether to plot intermediate fits}
}
\description{
Fit a cluster nearby peaks starting from a seed table of chemical shifts
}
