% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{fit_peaks}
\alias{fit_peaks}
\title{Fit peaks from a table of chemical shifts}
\usage{
fit_peaks(
  spec_list,
  cs_mat,
  fit_prev = NULL,
  spec_ord = 1:2,
  omega0_plus = c(0.075, 0.75),
  r2_start = 5,
  r2_bounds = c(0.5, 20),
  sc_start = NULL,
  sc_bounds = c(0, Inf),
  positive_only = TRUE,
  plot_fit_stages = TRUE
)
}
\arguments{
\item{spec_list}{list of spectra}

\item{cs_mat}{matrix of chemical shifts}

\item{fit_prev}{optional previous fit parameters}

\item{spec_ord}{order of spectral dimensions}

\item{omega0_plus}{omega0 bounds for each dimension}

\item{r2_start}{starting r2 values}

\item{r2_bounds}{numeric vector of length 2 with r2 bounds}

\item{sc_start}{starting scalar coupling values}

\item{sc_bounds}{numeric vector of length 2 with scalar coupling bounds}

\item{positive_only}{logical indicating whether to constrain m0 to be positive}

\item{plot_fit_stages}{logical indicating whether to plot intermediate fits}
}
\value{
A fitted parameter `list` (fit-output structure) containing fitted
  values in `fit_list` plus the original fit metadata and bounds.
}
\description{
Fit peaks from a table of chemical shifts
}
