% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{get_spec_peak_int}
\alias{get_spec_peak_int}
\title{Get spectra for individual peaks}
\usage{
get_spec_peak_int(
  fit_data,
  spec_type = c("start", "fit"),
  spec_idx = seq_along(fit_data$spec_data),
  peak_idx = seq_len(dim(fit_data$start_list$omega0)[2])
)
}
\arguments{
\item{fit_data}{fit_input or fit_output structure}

\item{spec_type}{character indicating which spectra to return}

\item{spec_idx}{indices of spectra to return}

\item{peak_idx}{indices of peaks to include}
}
\description{
Get spectra for individual peaks
}
