% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{height_assign}
\alias{height_assign}
\title{Calculate mapping from assigned peak list onto an unknown peak list}
\usage{
height_assign(assigned, unknown, thresh = 0.01)
}
\arguments{
\item{assigned}{two column matrix with assigned peak chemical shifts}

\item{unknown}{three column matrix with unknown peak chemical shifts and heights}

\item{thresh}{maximum distance (as a fraction of the two chemical shift ranges)}
}
\description{
This uses a greedy algorithm. It iterates over the unknown peaks in order of
decreasing height and assigns each unknown to the closest assigned peak within a
distance determined by the thresh parameter, as long as that peak wasn't already
assigned.
}
