% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{infer_acquisition_time}
\alias{infer_acquisition_time}
\title{Infer acquisition time for each dimension}
\usage{
infer_acquisition_time(fheader, empirically_correct = TRUE)
}
\arguments{
\item{fheader}{matrix of generalized ND parameters}

\item{empirically_correct}{logical indicating whether to apply empirical correction}
}
\description{
Infer acquisition time for each dimension
}
