% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{infer_aliasing}
\alias{infer_aliasing}
\title{Infer which dimension was directly acquired}
\usage{
infer_aliasing(fheader, phase_tolerance = 10)
}
\arguments{
\item{fheader}{matrix of generalized ND parameters}
}
\description{
Infer which dimension was directly acquired
}
