% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{infer_direct}
\alias{infer_direct}
\title{Infer which dimension was directly acquired}
\usage{
infer_direct(fheader)
}
\arguments{
\item{fheader}{matrix of generalized ND parameters}
}
\description{
Infer which dimension was directly acquired
}
