% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{infer_sweep_width}
\alias{infer_sweep_width}
\title{Infer original sweep width for each dimension}
\usage{
infer_sweep_width(fheader)
}
\arguments{
\item{fheader}{matrix of generalized ND parameters}
}
\description{
Infer original sweep width for each dimension
}
