% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{make_comb_list_omega0}
\alias{make_comb_list_omega0}
\title{Create list of data frames for}
\usage{
make_comb_list_omega0(omega0_name, sc_names, obs_mhz)
}
\arguments{
\item{single}{string (character vector of length 1) with scalar coupling names}
}
\description{
The only currently implemented way of splitting string is using whitespace
}
