% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{make_coupling_mat}
\alias{make_coupling_mat}
\title{Make a multiplet matrix with weights and scalar coupling coefficients}
\usage{
make_coupling_mat(sc_names)
}
\arguments{
\item{sc_names}{character vector with names of scalar couplings}
}
\description{
Make a multiplet matrix with weights and scalar coupling coefficients
}
