% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{nmr_pipe}
\alias{nmr_pipe}
\title{Process an FID with NMRPipe}
\usage{
nmr_pipe(
  in_path,
  out_path,
  ndim = 1,
  apod = NULL,
  sp = rbind(off = 0.5, end = 1, pow = 1, c = 0.5),
  zf = rbind(auto = ""),
  ps = rbind(p0 = 0, p1 = 0, di = ""),
  ext = NULL
)
}
\arguments{
\item{in_path}{input file path}

\item{out_path}{output file path}

\item{ndim}{number of dimensions}

\item{apod}{optional apodization argument matrix}

\item{sp}{optional sine-bell argument matrix}

\item{zf}{optional zero-fill argument matrix}

\item{ps}{optional phase correction argument matrix}

\item{ext}{optional extraction argument matrix}
}
\value{
The integer exit status from `system(...)` (typically `0` on success).
}
\description{
Process an FID with NMRPipe
}
