% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{nmrpipe_ft}
\alias{nmrpipe_ft}
\title{Fourier transform a 1D FID}
\usage{
nmrpipe_ft(fid)
}
\arguments{
\item{fid}{list with `int`, `header`, and `fheader` elements containing FID data}
}
\value{
A transformed spectrum `list` with updated `int`, `ppm`, `header`,
  and `fheader` fields.
}
\description{
This function aims to numerically match the NMRPipe `FT` function.
}
\details{
See the NMRPipe documentation: https://www.nmrscience.com/ref/nmrpipe/ft.html
}
\examples{
fid_path <- system.file("extdata", "noesy1d", "11", "test.fid", package = "fitnmr")
fid <- read_nmrpipe(fid_path, complex_data = TRUE)
sp <- nmrpipe_sp(fid)
zf <- nmrpipe_zf(sp)
ft <- nmrpipe_ft(zf)

}
