% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{nmrpipe_fti}
\alias{nmrpipe_fti}
\title{Inverse Fourier transform a 1D spectrum}
\usage{
nmrpipe_fti(ft)
}
\arguments{
\item{ft}{list with `int`, `header`, and `fheader` elements containing spectrum data}
}
\description{
This function aims to numerically match the NMRPipe `FT -inv` function.
}
\details{
See the NMRPipe documentation: https://www.nmrscience.com/ref/nmrpipe/ft.html
}
