% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{nmrpipe_ps}
\alias{nmrpipe_ps}
\title{Inverse Fourier transform a 1D spectrum}
\usage{
nmrpipe_ps(ft, p0 = 0, p1 = 0)
}
\arguments{
\item{ft}{list with `int`, `header`, and `fheader` elements containing spectrum data}

\item{p0}{equivalent to `-p0` flag}

\item{p1}{equivalent to `-p1` flag}
}
\value{
A modified spectrum `list` with phase-corrected `int` values and
  updated phase metadata in `header` and `fheader`.
}
\description{
This function aims to numerically match the NMRPipe `PS` function.
}
\details{
See the NMRPipe documentation: https://www.nmrscience.com/ref/nmrpipe/ps.html
}
