% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{nmrpipe_sp}
\alias{nmrpipe_sp}
\title{Apply sine-based window function to a 1D FID}
\usage{
nmrpipe_sp(fid, off = 0, end = 1, pow = 1, cval = 1, dmx = FALSE)
}
\arguments{
\item{fid}{list with `int`, `header`, and `fheader` elements containing FID data}

\item{off}{equivalent to `-off` flag}

\item{end}{equivalent to `-end` flag}

\item{pow}{equivalent to `-pow` flag}

\item{cval}{equivalent to `-c` flag}

\item{dmx}{equivalent to `-dmx` flag}
}
\value{
A modified FID `list` with updated `int`, `header`, and `fheader`
  values after sine-bell apodization.
}
\description{
This function aims to numerically match the NMRPipe `SP` function.
}
\details{
See the NMRPipe documentation: https://www.nmrscience.com/ref/nmrpipe/sp.html
}
\examples{
fid_path <- system.file("extdata", "noesy1d", "11", "test.fid", package = "fitnmr")
fid <- read_nmrpipe(fid_path, complex_data = TRUE)
sp <- nmrpipe_sp(fid)

}
