% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{nmrpipe_zf}
\alias{nmrpipe_zf}
\title{Apply zero filling to a 1D FID}
\usage{
nmrpipe_zf(fid, zf = 1, pad = NULL, size = NULL, auto = FALSE)
}
\arguments{
\item{fid}{list with `int`, `header`, and `fheader` elements containing FID data}

\item{zf}{equivalent to `-zf` flag}

\item{pad}{equivalent to `-pad` flag}

\item{size}{equivalent to `-size` flag}

\item{auto}{equivalent to `-auto` flag}
}
\description{
This function aims to numerically match the NMRPipe `ZF` function.
}
\details{
See the NMRPipe documentation: https://www.nmrscience.com/ref/nmrpipe/zf.html
}
