% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{omega0_comb_source_idx}
\alias{omega0_comb_source_idx}
\title{Get the first index in the omega0 array corresponding to each TRUE value in omega0_idx}
\usage{
omega0_comb_source_idx(param_list, omega0_idx = omega0_param_idx(param_list))
}
\description{
Get the first index in the omega0 array corresponding to each TRUE value in omega0_idx
}
