% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{omega0_param_idx}
\alias{omega0_param_idx}
\title{Get a list of logical arrays indicating which parameters correspond to peak positions}
\usage{
omega0_param_idx(
  param_list,
  dims = seq_len(dim(param_list[["group_list"]][["omega0"]])[1]),
  peaks = seq_len(dim(param_list[["group_list"]][["omega0"]])[2]),
  specs = seq_len(dim(param_list[["group_list"]][["omega0"]])[3])
)
}
\arguments{
\item{param_list}{parameter list}

\item{dims}{dimensions to include}

\item{peaks}{peaks to include}

\item{specs}{spectra to include}
}
\description{
Get a list of logical arrays indicating which parameters correspond to peak positions
}
