% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{param_array_to_comb_vec}
\alias{param_array_to_comb_vec}
\title{Determine vector of source parameters from destination array via least squares}
\usage{
param_array_to_comb_vec(
  param_array,
  comb_array,
  group_vec = NULL,
  resid_thresh = 1e-08
)
}
\arguments{
\item{param_array}{array of destination parameters}

\item{comb_array}{array of 2xN data frames with mapping between vector and array}

\item{group_vec}{optional vector of group numbers for source parameters}

\item{resid_thresh}{all residuals from least squares fit must be less than this value}
}
\description{
Determine vector of source parameters from destination array via least squares
}
