% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{param_list_to_arg_list}
\alias{param_list_to_arg_list}
\title{Convert a list of parameters for use with make_fit_input}
\usage{
param_list_to_arg_list(param_list)
}
\description{
Convert a list of parameters for use with make_fit_input
}
