% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{param_list_to_peak_df}
\alias{param_list_to_peak_df}
\title{Convert Fit to Data Frame}
\usage{
param_list_to_peak_df(param_list, spec_names = NULL)
}
\arguments{
\item{param_list}{list of fit parameters (or a list of such lists)}

\item{spec_names}{character vector of spectrum names}
}
\value{
A data frame with the following columns:
\describe{
  \item{\code{peak}}{peak number}
  \item{\code{fit}}{fit cluster number, with all peaks in the same cluster having the same \code{r2}, scalar couplings, and \code{m0}}
  \item{\code{f_pvalue}}{optional, p-value determined from F-test during iterative fitting}
  \item{\code{omega0_ppm_1}}{chemical shift of singlet/doublet center in the first dimension (ppm)}
  \item{\code{omega0_ppm_2}}{chemical shift of singlet/doublet center in the second dimension (ppm)}
  \item{\code{sc_hz_1}}{optional, scalar coupling of doublet in first dimension (Hz)}
  \item{\code{sc_hz_2}}{optional, scalar coupling of doublet in second dimension (Hz)}
  \item{\code{r2_hz_1}}{R2 in first dimension (Hz)}
  \item{\code{r2_hz_2}}{R2 in first dimension (Hz)}
  \item{...}{\code{m0} values for each spectrum}
}
}
\description{
Convert a parameter list into a peak data frame
}
\details{
This function takes the input or (if present) output parameters from a fit and converts them into a data frame. A parameter list must contain three lists:
\describe{
  \item{\code{start_list} or \code{fit_list}}{input or output values of the respective fit parameters}
  \item{\code{group_list}}{group numbers for the fit parameters}
  \item{\code{comb_list}}{coefficients for deriving fit parameters from a linear combination other auxiliary parameters}
}

This function currently assumes the fit parameters were generated by \code{\link{fit_peak_iter}}, \code{\link{fit_peak_cluster}}, or \code{\link{peak_df_to_param_list}}. These functions use a particular convention for \code{group_list} and \code{comb_list} to represent either singlets or doublets in each dimension of a 2D spectrum.

This function can take either a single parameter list or a list of parameter lists. If the latter is given, then the results from all the parameter lists will be combined into a single table.
}
