% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{param_list_to_tables_old}
\alias{param_list_to_tables_old}
\title{Convert a parameter list into a set of tables with resonance/nuclei/couplings}
\usage{
param_list_to_tables_old(param_list, tables)
}
\arguments{
\item{param_list}{param_list, fit_input, or fit_output structure}

\item{tables}{list with resonance/nuclei/couplings tables to update}
}
\description{
Convert a parameter list into a set of tables with resonance/nuclei/couplings
}
