% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{param_values}
\alias{param_values}
\alias{param_values<-}
\title{Get/set a subset of fitting parameters specified by a list of logical vectors}
\usage{
param_values(params, idx_list)

param_values(params, idx_list) <- value
}
\arguments{
\item{params}{a list of fitting parameters}

\item{idx_list}{a list with the same structure as \code{params} but with logical
vectors indicating which values should be return/set}
}
\value{
a vector of parameter values or the modified parameter list
}
\description{
Get/set a subset of fitting parameters specified by a list of logical vectors
}
