% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{peak_df_to_param_list}
\alias{peak_df_to_param_list}
\title{Convert a peak data frame to a parameter list}
\usage{
peak_df_to_param_list(peak_df, spectra)
}
\arguments{
\item{peak_df}{data frame of peak parameters}

\item{spectra}{list of spectra}
}
\value{
A parameter `list` with `start_list`, `group_list`, and `comb_list`
  inferred from `peak_df` and aligned to `spectra`.
}
\description{
Convert a peak data frame to a parameter list
}
