% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{perform_fit}
\alias{perform_fit}
\title{Perform a fit with an input data structure}
\usage{
perform_fit(
  fit_input,
  method = c("minpack.lm", "gslnls", "sparseLM", "L-BFGS-B"),
  ...
)
}
\arguments{
\item{fit_input}{fit_input structure}

\item{method}{fitting method}

\item{...}{additional arguments passed to the fitting routine}
}
\value{
The input `fit_input` list, augmented with fitted parameters in
  `fit_list` and fit diagnostics in `fit_rsstrace`, `fit_counts`, and
  `fit_time`.
}
\description{
Perform a fit with an input data structure
}
