% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{plot_fit_1d}
\alias{plot_fit_1d}
\title{Plot a one dimensional peak fit}
\usage{
plot_fit_1d(fit_data, always_show_start = FALSE)
}
\arguments{
\item{fit_data}{fit_input or fit_output structure}

\item{always_show_start}{logical indicating whether to show starting model when fit is present}
}
\value{
No return value, called for side effects (draws plots).
}
\description{
Plot a one dimensional peak fit
}
