% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{plot_fit_2d}
\alias{plot_fit_2d}
\title{Plot a two dimensional peak fit}
\usage{
plot_fit_2d(
  fit_output,
  spec_ord = seq_len(dim(fit_output$start_list$omega0)[1]),
  always_show_start = FALSE,
  main = NULL
)
}
\arguments{
\item{fit_output}{fit_output structure}

\item{spec_ord}{order of spectral dimensions to plot}

\item{always_show_start}{logical indicating whether to show starting model when fit is present}

\item{main}{optional title for the plot}
}
\description{
Plot a two dimensional peak fit
}
