% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{plot_resonances_1d}
\alias{plot_resonances_1d}
\title{Plot resonances from 1D fit}
\usage{
plot_resonances_1d(fit_data, always_show_start = FALSE, omega0_plus = 0.05)
}
\arguments{
\item{fit_data}{fit_input or fit_output structure}

\item{always_show_start}{show start parameters even if fit already done}

\item{omega0_plus}{length 3 vector giving ppm range for each dimension}
}
\description{
Plot resonances from 1D fit
}
