% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{plot_resonances_2d}
\alias{plot_resonances_2d}
\title{Plot resonances from 2D fit}
\usage{
plot_resonances_2d(
  fit_data,
  omega0_plus,
  resonances = unique(fit_data$resonance_names),
  low_frac = 0.05,
  field = TRUE,
  proj_frac = 0.2
)
}
\arguments{
\item{fit_data}{fit_input or fit_output structure}

\item{omega0_plus}{length 2 vector giving ppm range for each dimension}

\item{resonances}{character vector with resonances to plot}

\item{low_frac}{minimum absolute value (as a fraction of maximum intensity) at which to show contours}

\item{field}{logical indicating whether to show modeling of field inhomogeneity as separate peaks}

\item{proj_frac}{fraction of plot area reserved for 1D projections}
}
\value{
No return value, called for side effects (draws one plot layout per
  resonance).
}
\description{
Plot resonances from 2D fit
}
