% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{plot_sparse_1d}
\alias{plot_sparse_1d}
\title{Plot spectrum from 1D fit}
\usage{
plot_sparse_1d(
  fit_data,
  tables = NULL,
  spec_idx = 1,
  col_model = 2,
  col_resonance = NULL,
  lwd = 1,
  tick_spacing = 0.02,
  coupling_spacing = 0.01,
  coupling_marks = 0.009,
  xaxs = "i",
  yaxt = "n",
  bty = "n",
  always_show_start = FALSE,
  add = FALSE,
  ppm_map = make_map(get_spec_int(fit_data, "input", spec_idx)[[1]])
)
}
\arguments{
\item{fit_data}{fit_input or fit_output structure}

\item{tables}{list with resonances, nuclei, and couplings tables}

\item{spec_idx}{index of spectrum to plot}

\item{col_model}{color for modeled peak shapes}

\item{col_resonance}{colors for resonances}

\item{lwd}{line width}

\item{tick_spacing}{spacing between axis ticks (ppm)}

\item{coupling_spacing}{spacing between coupling annotations (fraction of plot)}

\item{coupling_marks}{size of coupling tick marks (fraction of plot)}

\item{xaxs}{x-axis style passed to \code{\link[graphics]{plot}}}

\item{yaxt}{y-axis style passed to \code{\link[graphics]{plot}}}

\item{bty}{box type passed to \code{\link[graphics]{plot}}}

\item{always_show_start}{logical indicating whether to show starting model when fit is present}

\item{add}{logical indicating whether to add to existing plot}

\item{ppm_map}{sparse axis map created by \code{\link{make_map}}}
}
\value{
No return value, called for side effects (draws a plot).
}
\description{
Plot spectrum from 1D fit
}
