% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{ppm_to_pts}
\alias{ppm_to_pts}
\title{Convert PPM values to points}
\usage{
ppm_to_pts(ppm_mat, fheader)
}
\arguments{
\item{ppm_mat}{matrix of ppm values}

\item{fheader}{matrix of generalized ND parameters}
}
\description{
Convert PPM values to points
}
