% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{read_resonance_tables}
\alias{read_resonance_tables}
\title{Read resonance, nuclei, coupling, and spin system tables}
\usage{
read_resonance_tables(
  resonances_file = "start_resonances.csv",
  nuclei_file = "start_nuclei.csv",
  couplings_file = "start_couplings.csv",
  comment.char = ""
)
}
\arguments{
\item{resonances_file}{path to resonances CSV file}

\item{nuclei_file}{path to nuclei CSV file}

\item{couplings_file}{path to couplings CSV file}

\item{comment.char}{comment character passed to \code{\link[utils]{read.csv}}}
}
\description{
Read resonance, nuclei, coupling, and spin system tables
}
