% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{resonance_to_param_list_old}
\alias{resonance_to_param_list_old}
\title{Convert data frame of resonances into a parameter list}
\usage{
resonance_to_param_list_old(spec, resonance, nuclei, couplings)
}
\arguments{
\item{spec}{single spectrum}

\item{resonance}{data frame with resonances}

\item{nuclei}{data frame with nuclei chemical shifts and R2 rates}

\item{data}{frame with scalar couplings}
}
\description{
Convert data frame of resonances into a parameter list
}
