% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{set_spinsystem_params}
\alias{set_spinsystem_params}
\title{Update HamiltonianMultiplet parameters from input data}
\usage{
set_spinsystem_params(
  spinsystems,
  omega0,
  omega0_comb,
  coupling_values = NULL,
  nucleus_names,
  ref_freq,
  spec_idx = 1
)
}
\arguments{
\item{spinsystems}{named list of HamiltonianMultiplet objects}

\item{omega0}{array of chemical shifts (ppm)}

\item{omega0_comb}{named vector of coupling values (Hz)}

\item{coupling_values}{named vector of coupling values (Hz)}

\item{nucleus_names}{array of nucleus names aligned with omega0}

\item{ref_freq}{numeric vector of reference frequencies (MHz)}

\item{spec_idx}{spectrum index to use for chemical shift conversions}
}
\description{
Update HamiltonianMultiplet parameters from input data
}
