% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{spec_bind}
\alias{spec_bind}
\title{Combine parameter lists referring to different spectra}
\usage{
spec_bind(...)
}
\arguments{
\item{...}{any number of parameter lists or a single list of parameter lists}
}
\description{
Combine parameter lists referring to different spectra
}
