% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{spec_overlap_mat}
\alias{spec_overlap_mat}
\title{Determine a matrix of fractional peak overlap}
\usage{
spec_overlap_mat(peak_int_list)
}
\arguments{
\item{peak_int_list}{list of peak intensity arrays}
}
\description{
Determine a matrix of fractional peak overlap
}
