% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{split_coupling_names}
\alias{split_coupling_names}
\title{Split string of scalar coupling names}
\usage{
split_coupling_names(name_char)
}
\arguments{
\item{single}{string (character vector of length 1) with scalar coupling names}
}
\description{
The only currently implemented way of splitting string is using whitespace
}
