% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{tables_to_param_list}
\alias{tables_to_param_list}
\title{Convert tables with resonance/nuclei/couplings to a parameter list}
\usage{
tables_to_param_list(spec_list, tables)
}
\arguments{
\item{spec_list}{list of spectra}

\item{tables}{list with resonance/nuclei/couplings tables}
}
\description{
Convert tables with resonance/nuclei/couplings to a parameter list
}
