% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{update_spinsystem_params}
\alias{update_spinsystem_params}
\title{Update HamiltonianMultiplet parameters from fit data}
\usage{
update_spinsystem_params(fit_data, spec_idx = 1, param_list = NULL)
}
\arguments{
\item{fit_data}{fit_input or fit_output structure}

\item{spec_idx}{spectrum index to use for chemical shift conversions}

\item{param_list}{optional parameter list (start_list or fit_list)}
}
\description{
Update HamiltonianMultiplet parameters from fit data
}
