% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{whz_to_pts}
\alias{whz_to_pts}
\title{Convert widths in Hz into points}
\usage{
whz_to_pts(whz_mat, fheader)
}
\arguments{
\item{whz_mat}{matrix of widths in Hz}

\item{fheader}{matrix of generalized ND parameters}
}
\description{
Convert widths in Hz into points
}
