% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{write_nmrdraw_peak_tab}
\alias{write_nmrdraw_peak_tab}
\title{Write an NMRDraw formatted peak table}
\usage{
write_nmrdraw_peak_tab(peak_tab, file_path)
}
\arguments{
\item{peak_tab}{data frame containing peak table data}

\item{file_path}{path to write the NMRDraw peak table file}
}
\value{
No return value, called for side effects (writes `file_path`).
}
\description{
Write an NMRDraw formatted peak table
}
