% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{collapse_na}
\alias{collapse_na}
\title{Collapse strings of repeated NAs in a vector with numeric names}
\usage{
collapse_na(x)
}
\arguments{
\item{x}{vector with names having numeric positional values}
}
\description{
Collapse strings of repeated NAs in a vector with numeric names
}
