% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{coupling_param_idx}
\alias{coupling_param_idx}
\title{Get a list of logical arrays indicating which parameters correspond to scalar couplings}
\usage{
coupling_param_idx(param_list, comb_idx_offset = 0)
}
\description{
Get a list of logical arrays indicating which parameters correspond to scalar couplings
}
