% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{fit_peak_cluster}
\alias{fit_peak_cluster}
\title{Fit a cluster nearby peaks starting from a seed table of chemical shifts}
\usage{
fit_peak_cluster(spec_list, cs_start, spec_ord, f_alpha_thresh = 0.001)
}
\description{
Fit a cluster nearby peaks starting from a seed table of chemical shifts
}
