% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{fit_peaks}
\alias{fit_peaks}
\title{Fit peaks from a table of chemical shifts}
\usage{
fit_peaks(spec_list, cs_mat, fit_prev = NULL, spec_ord = 1:2,
  omega0_plus = c(0.075, 0.75), r2_start = c(5, 5),
  positive_only = TRUE)
}
\description{
Fit peaks from a table of chemical shifts
}
