% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{limit_omega0_by_r2}
\alias{limit_omega0_by_r2}
\title{Add upper/lower limits based on the r2 value}
\usage{
limit_omega0_by_r2(fit_input, factor = 1.5)
}
\arguments{
\item{fit_input}{fit_input structure}

\item{factor}{multiplier applied to r2 to set omega0 bounds}
}
\value{
A modified `fit_input` list with updated `lower_list$omega0` and
  `upper_list$omega0` bounds.
}
\description{
Add upper/lower limits based on the r2 value
}
