% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{limit_omega0_by_r2}
\alias{limit_omega0_by_r2}
\title{Add upper/lower limits based on the r2 value}
\usage{
limit_omega0_by_r2(fit_input, factor = 1.5)
}
\description{
Add upper/lower limits based on the r2 value
}
