% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{make_map}
\alias{make_map}
\title{Create a sparse axis}
\usage{
make_map(
  x,
  max_spacing = 0.125,
  new_spacing = 0.025,
  starts = NULL,
  ends = NULL
)
}
\arguments{
\item{x}{numeric vector or array used to define the sparse axis}

\item{max_spacing}{maximum spacing between segments before inserting a gap}

\item{new_spacing}{spacing used for the inserted gap}

\item{starts}{optional numeric vector of segment start positions}

\item{ends}{optional numeric vector of segment end positions}
}
\description{
Create a sparse axis
}
