% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{nmrpipe_ft}
\alias{nmrpipe_ft}
\title{Fourier transform a 1D FID}
\usage{
nmrpipe_ft(fid)
}
\arguments{
\item{fid}{list with `int`, `header`, and `fheader` elements containing FID data}
}
\description{
This function aims to numerically match the NMRPipe `FT` function.
}
\details{
See the NMRPipe documentation: https://www.nmrscience.com/ref/nmrpipe/ft.html
}
