% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{nmrpipe_fti}
\alias{nmrpipe_fti}
\title{Inverse Fourier transform a 1D spectrum}
\usage{
nmrpipe_fti(ft)
}
\arguments{
\item{ft}{list with `int`, `header`, and `fheader` elements containing spectrum data}
}
\value{
A modified FID-like `list` with inverse-transformed `int` values and
  updated `header`/`fheader` metadata.
}
\description{
This function aims to numerically match the NMRPipe `FT -inv` function.
}
\details{
See the NMRPipe documentation: https://www.nmrscience.com/ref/nmrpipe/ft.html
}
\examples{
fid_path <- system.file("extdata", "noesy1d", "11", "test.fid", package = "fitnmr")
fid <- read_nmrpipe(fid_path, complex_data = TRUE)
sp <- nmrpipe_sp(fid)
zf <- nmrpipe_zf(sp)
ft <- nmrpipe_ft(zf)
fti <- nmrpipe_fti(ft)

}
