% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{noise_estimate}
\alias{noise_estimate}
\title{Estimate the noise in a spectrum}
\usage{
noise_estimate(
  x,
  height = TRUE,
  thresh = 10,
  plot_distributions = TRUE,
  peak_intensities = NULL,
  absolute = FALSE
)
}
\description{
Estimate the noise in a spectrum
}
