% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{param_list_to_peak_df}
\alias{param_list_to_peak_df}
\title{Convert a parameter list to a peak data frame}
\usage{
param_list_to_peak_df(param_list, spec_names = NULL)
}
\description{
Convert a parameter list to a peak data frame
}
