% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{peak_df_to_fit_input}
\alias{peak_df_to_fit_input}
\title{Convert a peak data frame to fit input}
\usage{
peak_df_to_fit_input(peak_df, spectra, ...)
}
\arguments{
\item{peak_df}{data frame of peak parameters}

\item{spectra}{list of spectra}

\item{...}{additional arguments passed to \code{\link{make_fit_input}}}
}
\value{
A `fit_input` list ready for optimization, as produced by
  `make_fit_input()`.
}
\description{
Convert a peak data frame to fit input
}
