% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{peak_df_to_fit_input}
\alias{peak_df_to_fit_input}
\title{Convert a peak data frame to fit input}
\usage{
peak_df_to_fit_input(peak_df, spectra, ...)
}
\description{
Convert a peak data frame to fit input
}
