% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{peak_df_to_param_list}
\alias{peak_df_to_param_list}
\title{Convert a peak data frame to a parameter list}
\usage{
peak_df_to_param_list(peak_df, spectra)
}
\description{
Convert a peak data frame to a parameter list
}
