% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{perform_fit}
\alias{perform_fit}
\title{Perform a fit with an input data structure}
\usage{
perform_fit(
  fit_input,
  method = c("minpack.lm", "gslnls", "sparseLM", "L-BFGS-B"),
  ...
)
}
\arguments{
\item{fit_input}{fit_input structure}

\item{method}{fitting method}

\item{...}{additional arguments passed to the fitting routine}
}
\value{
The input `fit_input` list, augmented with fitted parameters in
  `fit_list` and fit diagnostics in `fit_rsstrace`, `fit_counts`, and
  `fit_time`.
}
\description{
Perform a fit with an input data structure
}
\examples{
spec_file <- system.file("extdata", "t1", "1.ft2", package = "fitnmr")
spec <- read_nmrpipe(spec_file, dim_order = "hx")
fit_input <- make_fit_input(
  list(spec),
  omega0_start = matrix(c(8.5400, 119.76), nrow = 2),
  omega0_plus = c(0.075, 0.75),
  r2_start = 4,
  m0_start = 1e9
)
fit_output <- perform_fit(fit_input)
fit_output$start_list[1:3]
fit_output$fit_list[1:3]

}
