% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{perform_fit}
\alias{perform_fit}
\title{Perform a fit with an input data structure}
\usage{
perform_fit(
  fit_input,
  method = c("minpack.lm", "gslnls", "sparseLM", "L-BFGS-B"),
  ...
)
}
\description{
Perform a fit with an input data structure
}
