% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{plot_peak_df}
\alias{plot_peak_df}
\title{Plot a fits for series of spectra with parameters from a peak data frame}
\usage{
plot_peak_df(
  peak_df,
  spectra,
  noise_sigma = NULL,
  cex = 0.2,
  p0 = NULL,
  p1 = NULL
)
}
\description{
Plot a fits for series of spectra with parameters from a peak data frame
}
