% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonance.R
\name{plot_sparse_2d}
\alias{plot_sparse_2d}
\title{Plot spectrum from 2D fit}
\usage{
plot_sparse_2d(
  fit_data,
  tables = NULL,
  spec_idx = 1,
  col_model = 2,
  col_nucleus = NULL,
  lwd = 1,
  tick_spacing = 0.02,
  coupling_spacing = 0.01,
  coupling_marks = 0.009,
  xaxs = "i",
  yaxs = "i",
  low_frac = 0.05,
  bty = "n",
  always_show_start = FALSE,
  add = FALSE,
  ppm_map_list = NULL
)
}
\arguments{
\item{fit_data}{fit_input or fit_output structure}

\item{tables}{list with resonances, nuclei, and couplings tables}

\item{spec_idx}{index of spectrum to plot}

\item{col_model}{color for modeled peak shapes}

\item{col_nuclei}{colors for nuclei}
}
\description{
Plot spectrum from 2D fit
}
