% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrpipe.R
\name{read_nmrdraw_peak_tab}
\alias{read_nmrdraw_peak_tab}
\title{Read an NMRDraw formatted peak table}
\usage{
read_nmrdraw_peak_tab(file_path)
}
\arguments{
\item{file_path}{path to the NMRDraw peak table file}
}
\value{
A `data.frame` containing the peak table, with columns defined by the
  `VARS` line in the NMRDraw file.
}
\description{
Read an NMRDraw formatted peak table
}
