% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnmr.R
\name{update_fit_bounds}
\alias{update_fit_bounds}
\title{Update bounds on fitting parameters}
\usage{
update_fit_bounds(
  fit_input,
  omega0_r2_factor = NULL,
  r2_bounds = NULL,
  sc_bounds = NULL
)
}
\arguments{
\item{fit_input}{fit_input structure}

\item{omega0_r2_factor}{optional factor for omega0 bounds based on r2}

\item{r2_bounds}{optional numeric vector of length 2}

\item{sc_bounds}{optional numeric vector of length 2}
}
\description{
Update bounds on fitting parameters
}
